Code cleanup
 * Finalize key extractors - cannot be extended anyway
 * Allow to replace key prefix with predefined key extractor
 * Extract AbstractRpcTest methods to trait to avoid inheritance
 * Deprecate AbstractRpcTest
