+++ Decorators/CacheKeyStrategyInterface.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace ScayTrase\Api\Rpc\Decorators;
+
+use ScayTrase\Api\Rpc\RpcRequestInterface;
+
+interface CacheKeyStrategyInterface
+{
+    /**
+     * Create cache key for given RPC
+     *
+     * @param RpcRequestInterface $request
+     *
+     * @return string
+     */
+    public function getKey(RpcRequestInterface $request);
+}
