+++ composer.json
@@ -13,8 +13,8 @@
     "php": "~5.5|~7.0"
   },
   "require-dev": {
-    "phpunit/phpunit": "~4.5|~5.1",
-    "paragonie/random_compat": "~1.1@stable",
+    "phpunit/phpunit": "^4.8.35 | ^5.1 | ^6.0",
+    "paragonie/random_compat": "^1|^2",
     "psr/cache": "~1.0",
     "psr/log": "~1.0"
   },
